+++ Assets/Core/GameCore/GameCore.cs
@@ -1,3 +1,4 @@
+using UnityEngine;
 using System;
 using Cysharp.Threading.Tasks;
 using R3;
