+++ Assets/Core/GameCoreSystems/LoadingSplash/VgLoadingSplashManager.cs
@@ -34,9 +34,9 @@ namespace Core
 
         public void RegisterHooks(IGameCoreHookRegistry registry)
         {
-            registry.OnLoadStart(async _ => Show());
-            registry.OnLoadComplete(async _ => Hide());
-            registry.OnBootStart(async () => Show(), order: 0);
+            registry.OnLoadStart(_ => { Show(); return UniTask.CompletedTask; });
+            registry.OnLoadComplete(_ => { Hide(); return UniTask.CompletedTask; });
+            registry.OnBootStart(() => { Show(); return UniTask.CompletedTask; }, order: 0);
         }
 
         public void Init()
