+++ Assets/GameMain/Scripts/RunTime/GameCoreSystems/PlayerManager/PlayerManageSystem.cs
@@ -13,7 +13,7 @@ namespace GameMain.RunTime
         public void RegisterHooks(IGameCoreHookRegistry registry)
         {
             registry.OnInLevelEnter(async ctx => await SpawnPlayer(ctx));
-            registry.OnInLevelExit(async () => DespawnPlayer());
+            registry.OnInLevelExit(() => { DespawnPlayer(); return UniTask.CompletedTask; });
         }
 
         public async UniTask SpawnPlayer(LoadContext ctx)
